Move getMessageStats into helpers/discord, so I can reuse it
